requirements.txt
python-can
folium
scikit-learn
scipy
jupyter
osmnx
pyproj